Рефакторинг. Изменил config
