+++ app.js
@@ -2,15 +2,15 @@ const express = require("express");
 const mongoose = require("mongoose");
 const routers = require("./routes/index");
 
-const { DATABASE_URL, PORT, mongooseOptions } = require("./config/config");
+const { dataBase, port, mongooseOptions } = require("./config/config");
 
 const app = express();
 
-mongoose.connect(DATABASE_URL, mongooseOptions);
+mongoose.connect(dataBase, mongooseOptions);
 
 app.use(routers);
 
-app.listen(PORT, () => {
+app.listen(port, () => {
   // Если всё работает, консоль покажет, какой порт приложение слушает
-  console.info(`App listening on port ${PORT}`);
+  console.info(`App listening on port ${port}`);
 });
