+++ middlewares/error-logger.js
@@ -1,12 +1,6 @@
 const winston = require("winston");
 const expressWinston = require("express-winston");
 
-// логер запросов
-const requestLogger = expressWinston.logger({
-  transports: [new winston.transports.File({ filename: "request.log" })],
-  format: winston.format.json(),
-});
-
 // логгер ошибок
 const errorLogger = expressWinston.errorLogger({
   transports: [new winston.transports.File({ filename: "error.log" })],
@@ -14,6 +8,5 @@ const errorLogger = expressWinston.errorLogger({
 });
 
 module.exports = {
-  requestLogger,
   errorLogger,
 };
