+++ middlewares/request-logger.js
@@ -1,0 +1,12 @@
+const winston = require("winston");
+const expressWinston = require("express-winston");
+
+// логер запросов
+const requestLogger = expressWinston.logger({
+  transports: [new winston.transports.File({ filename: "request.log" })],
+  format: winston.format.json(),
+});
+
+module.exports = {
+  requestLogger,
+};
