+++ routes/index.js
@@ -1,18 +1,21 @@
 const { errors } = require("celebrate");
-const bodyParser = require("body-parser");
-const cookieParser = require("cookie-parser");
 const helmet = require("helmet");
 const router = require("express").Router();
+const bodyParser = require("body-parser");
+const cookieParser = require("cookie-parser");
 
-const registration = require("./registration");
-const login = require("./login");
-const users = require("./users");
-const articles = require("./articles");
-const pathInvalid = require("./path-invalid ");
-const auth = require("../middlewares/auth");
-const errorHandler = require("../middlewares/error-handler");
-const { requestLogger, errorLogger } = require("../middlewares/logger");
 const limiter = require("../middlewares/limiter");
+const { requestLogger } = require("../middlewares/request-logger");
+const registration = require("./registration-user");
+const login = require("./login-user");
+const auth = require("../middlewares/auth");
+const readUser = require("./read-user");
+const creatAarticle = require("./create-article");
+const readArticles = require("./read-articles");
+const deleteArticle = require("./delete-article");
+const pathInvalid = require("./path-invalid ");
+const { errorLogger } = require("../middlewares/error-logger");
+const errorHandler = require("../middlewares/error-handler");
 
 router.use(helmet());
 router.use(bodyParser.json());
@@ -20,11 +23,14 @@ router.use(cookieParser());
 
 router.use(limiter);
 router.use(requestLogger);
+
 router.use("/signup", registration);
 router.use("/signin", login);
 router.use(auth);
-router.use("/users", users);
-router.use("/articles", articles);
+router.use("/users", readUser);
+router.use("/articles", creatAarticle);
+router.use("/articles", readArticles);
+router.use("/articles", deleteArticle);
 router.use("*", pathInvalid);
 
 router.use(errorLogger);
