+++ routes/login-user.js
@@ -1,8 +1,8 @@
 const router = require("express").Router();
 
 const { loginCheckSchema } = require("../validation/user-validation-celebrait");
-const { login } = require("../controllers/users");
+const { loginUser } = require("../controllers/login-user");
 
-router.post("/", loginCheckSchema, login);
+router.post("/", loginCheckSchema, loginUser);
 
 module.exports = router;
