+++ routes/read-articles.js
@@ -1,0 +1,7 @@
+const router = require("express").Router();
+
+const { readArticle } = require("../controllers/read-articles");
+
+router.get("/", readArticle);
+
+module.exports = router;
