+++ routes/registration-user.js
@@ -3,7 +3,7 @@ const router = require("express").Router();
 const {
   registrationCheckSchema,
 } = require("../validation/user-validation-celebrait");
-const { createUser } = require("../controllers/users");
+const { createUser } = require("../controllers/create-user");
 
 router.post("/", registrationCheckSchema, createUser);
